+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 1. [](#new)
     * Made `Grav::redirect()` to accept `Route` class
+    * Added `translated()` method to `PageTranslateInterface`
 2. [](#improved)
    * Upgraded vendor libs for PHP 8.1 compatibility
    * Upgraded to **composer v2.1.14** for PHP 8.1 compatibility